# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "chat_API",
      "contact": {},
      "version": "1.0"
    },
    "paths": {
      "/api/v1/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "register",
          "operationId": "register",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/registerrequest"
                    },
                    {
                      "example": {
                        "firstname": "ofe",
                        "lastname": "rayo",
                        "email": "ofe@hotmail.com",
                        "password": "1234567aA@"
                      }
                    }
                  ]
                },
                "example": {
                  "firstname": "ofe",
                  "lastname": "rayo",
                  "email": "ofe@hotmail.com",
                  "password": "1234567aA@"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        },
        "get": {
            "security": [
                {
                  "bearer": []
                }
              ],
          "tags": [
            "users"
          ],
          "summary": "get all users",
          "operationId": "getallusers",
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
      "/api/v1/users/login": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "login user",
          "operationId": "loginuser",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/loginuserrequest"
                    },
                    {
                      "example": {
                        "email": "isa_more7@hotmail.com",
                        "password": "1234567aA@"
                      }
                    }
                  ]
                },
                "example": {
                  "email": "isa_more7@hotmail.com",
                  "password": "1234567aA@"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
      "/api/v1/conversations": {
        "post": {
            "security": [
                {
                  "bearer": []
                }
              ],
          "tags": [
            "conversations"
          ],
          "summary": "create conversation",
          "operationId": "createconversation",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/createconversationrequest"
                    },
                    {
                      "example": {
                        "userId": 1,
                        "participantId": 2
                      }
                    }
                  ]
                },
                "example": {
                  "userId": 1,
                  "participantId": 2
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
      "/api/v1/conversations/group": {
        "post": {
            "security": [
                {
                  "bearer": []
                }
              ],
          "tags": [
            "conversations"
          ],
          "summary": "creat group",
          "operationId": "creatgroup",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/creatgrouprequest"
                    },
                    {
                      "example": {
                        "userId": 1,
                        "participantsIds": [
                          3,
                          4,
                          5
                        ],
                        "title": "family"
                      }
                    }
                  ]
                },
                "example": {
                  "userId": 1,
                  "participantsIds": [
                    3,
                    4,
                    5
                  ],
                  "title": "family"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
      "/api/v1/conversations/{id}": {
        "get": {
          "tags": [
            "conversations"
          ],
          "summary": "get all conversations by user",
          "operationId": "getallconversationsbyuser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/messages/conversation/{id}": {
        "post": {
          "tags": [
            "messages"
          ],
          "summary": "create message",
          "operationId": "createmessage",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 6
              }
            }
          ],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/createmessagerequest"
                    },
                    {
                      "example": {
                        "content": "que paso?",
                        "senderId": 2
                      }
                    }
                  ]
                },
                "example": {
                  "content": "que paso?",
                  "senderId": 2
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "tags": [
            "messages"
          ],
          "summary": "get conversation messages",
          "operationId": "getconversationmessages",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/users/{id}": {
        "put": {
            "security": [
                {
                  "bearer": []
                }
              ],
          "tags": [
            "messages"
          ],
          "summary": "upload image",
          "operationId": "uploadimage",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 6
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "encoding": {},
                "schema": {
                  "required": [
                    "avatar"
                  ],
                  "type": "object",
                  "properties": {
                    "avatar": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            },
            "required": false
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      }
    },
    "components": {
      "schemas": {
        "registerrequest": {
          "title": "registerrequest",
          "required": [
            "firstname",
            "lastname",
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string"
            },
            "lastname": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "firstname": "ofe",
            "lastname": "rayo",
            "email": "ofe@hotmail.com",
            "password": "1234567aA@"
          }
        },
        "loginuserrequest": {
          "title": "loginuserrequest",
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "email": "isa_more7@hotmail.com",
            "password": "1234567aA@"
          }
        },
        "createconversationrequest": {
          "title": "createconversationrequest",
          "required": [
            "userId",
            "participantId"
          ],
          "type": "object",
          "properties": {
            "userId": {
              "type": "integer",
              "format": "int32"
            },
            "participantId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "example": {
            "userId": 1,
            "participantId": 2
          }
        },
        "creatgrouprequest": {
          "title": "creatgrouprequest",
          "required": [
            "userId",
            "participantsIds",
            "title"
          ],
          "type": "object",
          "properties": {
            "userId": {
              "type": "integer",
              "format": "int32"
            },
            "participantsIds": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "description": ""
            },
            "title": {
              "type": "string"
            }
          },
          "example": {
            "userId": 1,
            "participantsIds": [
              3,
              4,
              5
            ],
            "title": "family"
          }
        },
        "createmessagerequest": {
          "title": "createmessagerequest",
          "required": [
            "content",
            "senderId"
          ],
          "type": "object",
          "properties": {
            "content": {
              "type": "string"
            },
            "senderId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "example": {
            "content": "que paso?",
            "senderId": 2
          }
        }
      },
      "securitySchemes": {
        "bearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
    "tags": [
      {
        "name": "users"
      },
      {
        "name": "conversations"
      },
      {
        "name": "messages"
      }
    ]
  }